requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
pillow>=8.3.1
scikit-image>=0.18.3
tensorflow>=2.8.0
torch>=1.9.0
torchvision>=0.10.0
gradio>=4.0.0 